requirements.txt
boto3==1.24.92
botocore==1.27.92
channels==3.0.5
confluent-kafka==1.9.2
Django==3.2.16
django-ckeditor==6.3.2
django-cors-headers==3.13.0
web3==5.31.3
django-environ==0.9.0
django-redis==5.2.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djangorestframework-api-response==0.0.30
gunicorn==20.1.0
Pillow==9.2.0
mutagen==1.46.0
requests==2.28.1
psycopg2==2.9.4
python-slugify==6.1.2
PyJWT==2.5.0
whitenoise==6.2.0
uvicorn==0.20.0
jsonpickle==3.0.1
rsa==4.9
blockfrost-python==0.5.3